Send Request To Upgrade Role

+++ src/main/java/com/hcmus/auction/service/impl/UserServiceImpl.java
@@ -1,26 +1,25 @@
 package com.hcmus.auction.service.impl;
 
+import com.hcmus.auction.common.variable.ErrorMessage;
+import com.hcmus.auction.exception.GenericException;
 import com.hcmus.auction.model.dto.UserDTO;
 import com.hcmus.auction.model.entity.User;
 import com.hcmus.auction.model.mapper.UserMapper;
 import com.hcmus.auction.repository.UserRepository;
 import com.hcmus.auction.service.definition.GenericService;
 import com.hcmus.auction.service.definition.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class UserServiceImpl implements GenericService<UserDTO, String>, UserService {
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private UserMapper userMapper;
-
-    @Autowired
-    private FavoriteProductServiceImpl favoriteProductService;
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
+    private final FavoriteProductServiceImpl favoriteProductService;
+    private final RoleHistoryServiceImpl roleHistoryService;
 
     @Override
     public UserDTO getById(String userId) {
@@ -30,11 +29,22 @@ public class UserServiceImpl implements GenericService<UserDTO, String>, UserSer
 
     @Override
     public void addNewFavoriteProduct(String userId, String productId) {
+        if (this.getById(userId) == null)
+            throw new GenericException(ErrorMessage.NOT_EXISTED_USER.getMessage());
         favoriteProductService.addNewFavoriteProduct(userId, productId);
     }
 
     @Override
     public void deleteFavoriteProduct(String userId, String productId) {
+        if (this.getById(userId) == null)
+            throw new GenericException(ErrorMessage.NOT_EXISTED_USER.getMessage());
         favoriteProductService.deleteFavoriteProduct(userId, productId);
+    }
+
+    @Override
+    public void addNewRoleHistory(String userId) {
+        if (this.getById(userId) == null)
+            throw new GenericException(ErrorMessage.NOT_EXISTED_USER.getMessage());
+        roleHistoryService.addNewRoleHistory(userId);
     }
 }
